Changed floats to map to Firebolt DOUBLE types.

+++ dbt/adapters/firebolt/impl.py
@@ -223,7 +223,7 @@ class FireboltAdapter(SQLAdapter):
         types = {
             'str': 'TEXT',
             'int': 'LONG',
-            'float': 'FLOAT',
+            'float': 'DOUBLE',
             'date': 'DATE',
             'datetime': 'DATE',
             'bool': 'BOOLEAN',
